requirements.txt
# 核心依赖
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.21.0
scipy>=1.7.0

# 强化学习环境
gymnasium>=0.28.0
gym>=0.21.0

# 数据处理与可视化
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.3.0
pillow>=9.0.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8

# 配置管理
pyyaml>=6.0
omegaconf>=2.3.0

# 日志与监控
tensorboard>=2.12.0
wandb>=0.15.0
tqdm>=4.65.0

# 图神经网络（用于基线算法）
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# 工具库
colorama>=0.4.6
tabulate>=0.9.0
scikit-learn>=1.0.0

# 代码质量
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0

# 其他
psutil>=5.9.0

